backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==4.2.0
psycopg2-binary==2.9.10
Pillow==10.4.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
python-dotenv==1.0.0
gunicorn==21.2.0